requirements.txt
absl-py==1.4.0
albumentations==1.3.1
altgraph @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot11/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-141/altgraph-0.17.2-py2.py3-none-any.whl
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
appnope==0.1.3
asgiref==3.3.4
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
beautifulsoup4==4.12.2
black==23.10.1
blinker==1.6.2
blis==0.7.11
bs4==0.0.2
cached-property==1.5.2
cachelib==0.10.2
cachetools==5.3.1
catalogue==2.0.10
certifi==2020.12.5
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
cloudpathlib==0.16.0
cloudpickle==3.0.0
colorama==0.4.4
comm==0.1.3
commonmark==0.9.1
confection==0.1.4
contourpy==1.1.0
cryptography==41.0.7
cycler==0.11.0
cymem==2.0.8
debugpy==1.6.7
decorator==5.1.1
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@898507047cf441a1e4be7a729270961c401c4354
dictdiffer==0.9.0
efficientnet-pytorch==0.7.1
exceptiongroup==1.1.2
executing==1.2.0
filelock==3.12.2
Flask==2.3.2
Flask-Cors==4.0.0
Flask-OAuthlib==0.9.6
Flask-ReCaptcha==0.4.2
Flask-Toastr==0.5.8
Flask-WTF==1.2.1
flatbuffers==23.5.26
fonttools==4.41.1
fsspec==2023.10.0
future @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot11/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-141/future-0.18.2-py3-none-any.whl
fvcore==0.1.5.post20221221
gast==0.4.0
gdown==4.7.1
geojson==3.0.1
gitdb==4.0.10
GitPython==3.1.32
gmaps==0.9.0
google-api-core==2.12.0
google-api-python-client==2.103.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.60.0
goto-statement==1.2
grpcio==1.56.2
h11==0.14.0
h5py==3.9.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.26.0
huggingface-hub==0.18.0
hydra-core==1.3.2
idna==3.4
imageio==2.31.6
importlib-metadata==6.8.0
importlib-resources==6.0.0
iniconfig==2.0.0
iopath==0.1.9
ipykernel==6.24.0
ipython==8.14.0
ipywidgets==8.1.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyterlab-widgets==3.0.8
keras==2.13.1
kiwisolver==1.4.4
langcodes==3.3.0
lazy_loader==0.3
libclang==16.0.6
lightning-utilities==0.9.0
macholib @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot11/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-141/macholib-1.15.2-py2.py3-none-any.whl
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mpmath==1.3.0
munch==4.0.0
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
notion==0.0.28
notion-client==2.2.1
notion-database==1.1.0
numpy==1.24.3
oauthlib==2.1.0
omegaconf==2.3.0
opencv-python==4.8.0.74
opencv-python-headless==4.8.1.78
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.9.1
pluggy==1.2.0
portalocker==2.8.2
preshed==3.0.9
pretrainedmodels==0.7.4
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.15.1
PyJWT==2.8.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.4.0
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.0
QtPy==1.9.0
qudida==0.0.4
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.1.0
rsa==4.9
safetensors==0.4.0
scikit-image==0.22.0
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
segmentation-models-pytorch==0.3.3
six @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot11/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-141/six-1.15.0-py2.py3-none-any.whl
smart-open==6.4.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.2
sympy==1.12
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-macos==2.13.0
termcolor==2.3.0
text-unidecode==1.3
thinc==8.2.1
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.9.26
timm==0.9.2
tokenizers==0.15.0
tomli==2.0.1
torch==2.0.0
torchdata==0.7.1
torchmetrics==1.0.1
torchsummary==1.5.1
torchtext==0.16.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.35.2
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
ultralytics==8.0.203
uritemplate==4.1.1
urllib3==1.26.16
wasabi==1.1.2
wcwidth==0.2.6
weasel==0.3.4
Werkzeug==2.3.6
widgetsnbextension==4.0.8
wrapt==1.15.0
WTForms==3.1.1
yacs==0.1.8
zipp==3.16.2
